feat(activity): 引入MQ处理活动SKU库存一致性

+++ market-forge-domain/src/main/java/com/lavyoung/marketforge/domain/activity/model/entity/ActivitySkuEntity.java
@@ -18,6 +18,6 @@ public record ActivitySkuEntity(
         Long sku,
         Long activityId,
         Long activityCountId,
-        Long stockCount
+        Integer stockCount
 ) {
 }
